wsyn change voice per track implementation
instead of sending a VOX command when a note triggers, and not using a cv callback, make it just send a VEL command in the trigger callback and a PITCH command in the cv callback. This change is fairly meaningless without the ability to update the CV without a trigger, which is not possible in vanilla ansible. Next commit will add this functionality