create new per-track setting 'stream_notes'
This is similar to the teletype and trigger clocking setting, and it is toggled via the next column adjacent to them (column 3, y=2). With it enabled, the track's CV output is updated every time the note changes, not just when a trigger fires. This enables ex. firing a trigger, then changing the pitch of the voice while the envelope decays without retriggering the envelope