add a third mode for txo, non-enveloped oscillator
This mode uses only the CV outputs of txo, generating a simple non-enveloped oscillation. Pitch changes with the CV output. It is intended for the oscillator being fed through an external filter/amplifier/envelope, where the gate for the envelope comes from Ansible's TR output, leaving the corresponding TR output on TXO free for other uses.